requirements.txt
customtkinter
pillow
tksheet
